requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
kombu==5.3.7
prompt_toolkit==3.0.47
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.2
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
